requirements.txt
pandas
numpy
scikit-learn
mysql-connector-python
python-dotenv


#-e .


